pyproject.toml
[project]
name = "ibm-cloud-token-mcp"
version = "0.1.0"
description = "MCP server for IBM Cloud Bearer Token management"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "IBM Cloud Token MCP", email = "noreply@example.com"}
]
dependencies = [
    "fastmcp>=0.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ibm-cloud-token-mcp = "ibm_cloud_token_mcp.__main__:main"

[project.urls]
Homepage = "https://github.com/yourusername/ibm-cloud-token-mcp"
Repository = "https://github.com/yourusername/ibm-cloud-token-mcp"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ibm_cloud_token_mcp"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "respx>=0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]